pyproject.toml
[tool.poetry]
name = "import-bank-details"
version = "0.1.0"
description = "Python script to import, merge, and export different bank statements together."
authors = ["Giampaolo Casolla <giampaolocasolla@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.2.2"
PyYAML = "^6.0.2"
openpyxl = "^3.1.5"
pydantic = "^2.8.2"
python-dotenv = "^1.0.1"
openai = "^1.52.0"
duckduckgo-search = "^7.5.0"

[tool.poetry.dev-dependencies]
ipykernel = "^6.29.5"
nbformat = "^5.10.4"
pytest = "^6.2.4"
black = "^24.4.2"
flake8 = "^3.9.2"
isort = "^5.10.1"
mypy = "^1.2.0"
pre-commit = "^3.0.0"
types-pyyaml = "^6.0.12.20240917"
pytest-cov = "^4.1.0"

[tool.isort]
profile = "black"
line_length = 130
skip_gitignore = true

[tool.black]
line-length = 130

[tool.mypy]
python_version = "3.11"
warn_unused_configs = false
disallow_untyped_defs = false
warn_return_any = false
ignore_missing_imports = true
check_untyped_defs = false

[tool.pytest.ini_options]
markers = [
    "integration: marks tests as integration tests",
]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"

[tool.coverage.run]
source = ["import_bank_details"]
omit = ["tests/*", "**/__init__.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "pass",
    "raise ImportError",
]
fail_under = 80

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
